requirements.txt
pymongo
python-dotenv
httpx
mutmut
pytest
hypothesis
google.generativeai
tenacity
jsonschema
func-timeout